replace legacy splitting code with new splitting code
