--- src/ophys_etl/modules/mesoscope_splitting_2022/__init__.py
@@ -1,1 +1,0 @@
-# empty
